feat: 지수 정보 연동기능 구현

+++ src/main/java/com/part2/findex/indexinfo/dto/response/IndexInfoDto.java
@@ -7,7 +7,7 @@ public record IndexInfoDto(
     Long id,
     String indexClassification,
     String indexName,
-    Integer employedItemsCount,
+    Double employedItemsCount,
     String basePointInTime,
     Double baseIndex,
     String sourceType,
